+++ include/hdlc_parser/writer/pcap/header.hpp
@@ -16,8 +16,8 @@ struct PcapGlobalHeader {
 struct PcapPacketHeader {
     uint32_t ts_sec;
     uint32_t ts_usec;
-    uint32_t incl_len;
-    uint32_t orig_len;
+    size_t incl_len;
+    size_t orig_len;
 };
 
 #pragma pack(pop)
